+++ infra/postgres/migrations/0016_workflow_planning.down.sql
@@ -1,0 +1,5 @@
+BEGIN;
+DROP TABLE IF EXISTS workflow_plan_attempt;
+DROP TABLE IF EXISTS workflow_plan;
+DELETE FROM schema_migration WHERE version='0016_workflow_planning';
+COMMIT;